1st prototype of the discord presence Refined the requirements.txt Made gen_summary.py more resilient to missing log files Moved find_song_rating to the sdvx_utils.py Slight improvements on build.cmd

+++ build.cmd
@@ -6,7 +6,8 @@ set target_ocr=ocr_reporter
 set target_manager=manage_score
 set target_updater=update
 set target_sync=play_log_sync
-set pyin="%userprofile%\AppData\Roaming\Python\Python313\Scripts\pyinstaller.exe"
+set pyin="%userprofile%\AppData\Local\Programs\Python\Python314\Scripts\pyinstaller.exe"
+REM set pyin="%userprofile%\AppData\Roaming\Python\Python313\Scripts\pyinstaller.exe"
 
 set SDVX=Not Build
 set OCR=Not Build
@@ -80,7 +81,7 @@ if "%MANAGER%" == "Build" %pyin% %target_manager%.py --clean --noconsole --onefi
 if "%UPDATER%" == "Build" %pyin% %target_updater%.py --clean --noconsole --onefile --icon=icon.ico --add-data "icon.ico;./" --hidden-import=tkinter --hidden-import=tkinter.filedialog -y
 if "%SYNC%" == "Build" %pyin% %target_sync%.py --clean --noconsole --onefile --icon=icon.ico --add-data "icon.ico;./" --hidden-import=tkinter --hidden-import=tkinter.filedialog -y
 
-powershell Remove-Item -Path %target_sdvx% -Recurse -Force
+if exist %target_sdvx% powershell Remove-Item -Path %target_sdvx% -Recurse -Force
 
 REM binaries
 xcopy /Y dist\*.exe %target_sdvx%\
@@ -99,5 +100,16 @@ del /F /Q %target_sdvx%\out\rival*.pkl
 
 REM Zip everything
 if exist %target_sdvx%_en_all.zip del /F /Q %target_sdvx%_en_all.zip
-powershell -Command "Compress-Archive -Path %target_sdvx%\* -DestinationPath %target_sdvx%_en_all.zip"
+
+if "%SDVX%" == "Build" (
+    if "%OCR%" == "Build" (
+        if "%MANAGER%" == "Build" (
+            if "%UPDATER%" == "Build" (
+                if "%SYNC%" == "Build" (
+                    powershell -Command "Compress-Archive -Path %target_sdvx%\* -DestinationPath %target_sdvx%_en_all.zip"
+                )
+            )
+        )
+    )
+)
